requirements.txt
absl-py==2.3.1
ale-py==0.11.2
astunparse==1.6.3
attrs==25.4.0
box2d-py==2.3.5
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
dm-tree==0.1.9
Farama-Notifications==0.0.4
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
gymnasium==1.1.1
h5py==3.15.1
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
lz4==4.4.4
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpe2==0.0.1
mpmath==1.3.0
msgpack==1.1.2
namex==0.1.0
networkx==3.5
numpy==2.2.6
onnx==1.19.1
onnxruntime==1.23.2
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
ormsgpack==1.7.0
packaging==25.0
pandas==2.3.3
pettingzoo==1.25.0
pillow==12.0.0
protobuf==6.33.0
psutil==7.1.1
pyarrow==21.0.0
pygame==2.6.1
pyglet==1.5.27
Pygments==2.19.2
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
ray==2.50.1
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
scipy==1.16.2
six==1.17.0
stable_baselines3==2.7.0
swig==4.3.1.post0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tensorflow==2.20.0
termcolor==3.1.0
torch==2.9.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==2.0.0
